pyproject.toml
[project]
name = "youtube-extract"
version = "0.2.0"
description = "Open Heavens Church — extrai transcrições do YouTube, segmenta o culto com IA, gera resumo pastoral e guia Open Groups com identidade visual"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "youtube-transcript-api>=1.0.0",
    "anthropic>=0.40.0",
    "python-dotenv>=1.0.0",
    "yt-dlp>=2024.8.0",
    "jinja2>=3.1.0",
    "markdown>=3.5.0",
    "weasyprint>=62.0",
    "rich>=13.7.0",
    "typer>=0.12.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    "ruff>=0.5.0",
    "black>=24.0",
    "mypy>=1.10",
    "pre-commit>=3.7",
]

[project.scripts]
youtube-extract = "youtube_extract.app:extract_entry"
youtube-summarize = "youtube_extract.app:summarize_entry"
youtube-prepare-claude = "youtube_extract.app:prepare_claude_entry"
youtube-ministracao = "youtube_extract.app:ministracao_entry"
open-heavens = "youtube_extract.app:app"

[build-system]
requires = ["setuptools>=61"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
youtube_extract = [
    "templates/*.j2",
    "templates/prompts/*.j2",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]
ignore = ["E501", "B008", "SIM102", "SIM105", "SIM108", "UP007", "UP035", "UP045"]

[tool.black]
line-length = 100
target-version = ["py310", "py311", "py312"]

[tool.mypy]
python_version = "3.10"
warn_unused_ignores = true
warn_redundant_casts = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q"
